refactor(trusty-review): split subprocess_analyze_client to satisfy 500-line cap (refs #632)
The flat 599-line subprocess_analyze_client.rs exceeded the hard 500-line
cap enforced by scripts/check_line_cap.sh and the CI line-cap workflow,
causing PR #634 to fail.

Split into a focused module directory:
- subprocess_analyze_client/mod.rs  (126 lines) — wire-format types,
  map_report() helper, constants, and re-export
- subprocess_analyze_client/client.rs (274 lines) — SubprocessAnalyzeClient
  struct, all trait impl methods, spawn_analyze_review() sync helper
- subprocess_analyze_client/tests.rs (226 lines) — all 11 unit tests
  (isolated under #[cfg(test)] mod tests)

No allowlist entry added; all new files are ≤ 500 lines. Zero functional
changes — the public API (SubprocessAnalyzeClient re-exported from
integrations/mod.rs) is identical.
